Huge progress. All changes and implementations listed below.
First off, AWS S3 has been successfully implemented into the application. When the page gets deployed, there will be a documentation page for all methods associated with AWS S3 and file uploading.

Secondly, as mentioned in the previous paragraph, a documentation page has been added and reviewed by peers. The page will be for contributors and even for future collaborators to understand the scope of the project and the methods associated with it. This is to prevent confusion and to help encourage understanding and productivity.

Lastly, tiny adjustments have been made to the main page to allow only 3 files to be uploaded at once. This is to prevent too many files to be uploaded into the buckets.

Personal note: I would like to work on the file retrieval part of the project. After that, it can finally be deployed using a deployment service of my choice, perhaps AWS. Then build automation and automated testing can finally be implemented as the final piece to the project.